[CHORE] Unblock PropertiesConfigEC2Version.kt and Block PropertiesConfigLocalVersion.kt
